Trying to debug code, had to fix some issues
Two big issues fixed. All in N6_RJ script.

Removed the porosity on the attachment term because it actually should never have been there (there was a hidden porosity term on the attachment term).

Also fixed another error in the derivation where x^[(c-1)(b-1)] was fixed to x^(c+b-2)

Also found a small error in the y-vector call for the jacobian for the mobile interior nodes with respect to the associated attached node (called the attached ndoe value isntead of the mobile node value)